+++ src/main/java/roomescape/auth/application/AuthService.java
@@ -3,7 +3,6 @@ package roomescape.auth.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.auth.dto.AuthenticatedMember;
-import roomescape.auth.infrastructure.TokenService;
 import roomescape.global.exception.AuthenticationException;
 import roomescape.member.model.Member;
 import roomescape.member.model.MemberRepository;
